Create syncer interface
